milbot/requirements.txt
certifi
python-dotenv
redis
requests
slackclient
